remove asOf and combine var models
