+++ models/credmark/algorithms/value_at_risk/dto.py
@@ -1,4 +1,3 @@
-from datetime import date
 from typing import List
 
 from credmark.dto import (
@@ -18,7 +17,6 @@ from credmark.cmf.types import (
 class HistoricalPriceInput(DTO):
     token: Token
     window: str  # e.g. '30 day'
-    asOf: date
 
 
 class VaRHistoricalInput(IterableListGenericDTO[PriceList]):
@@ -30,7 +28,6 @@ class VaRHistoricalInput(IterableListGenericDTO[PriceList]):
 
 
 class ContractVaRInput(DTO):
-    asOf: date
     window: str
     interval: int
     confidences: List[float]
@@ -38,8 +35,7 @@ class ContractVaRInput(DTO):
     class Config:
         schema_extra = {
             'examples': [
-                {'asOf': '2022-02-17',
-                 'window': '2 days',
+                {'window': '2 days',
                  'interval': 1,
                  'confidences': [0.01, 0.05]
                  }]
