+++ models/credmark/algorithms/value_at_risk/var_aave.py
@@ -1,20 +1,15 @@
-import json
-from datetime import datetime, timezone
-
 from credmark.cmf.model import Model
 from credmark.dto import EmptyInput
 
 from credmark.cmf.types import (
     Position,
     Portfolio,
-    BlockNumber,
     PriceList,
 )
 
 from models.credmark.algorithms.value_at_risk.dto import (
-    HistoricalPriceInput,
-    VaRHistoricalInput,
     ContractVaRInput,
+    PortfolioVaRInput,
 )
 
 
@@ -24,7 +19,7 @@ from models.credmark.protocols.lending.aave.aave_v2 import (
 
 
 @Model.describe(slug="finance.var-aave",
-                version="1.0",
+                version="1.1",
                 display_name="Aave V2 VaR",
                 description="Calcualte the VaR of Aave contract of its net asset",
                 input=ContractVaRInput,
@@ -43,13 +38,9 @@ class AaveV2GetVAR(Model):
     """
 
     def run(self, input: ContractVaRInput) -> dict:
-        asof_dt = datetime.combine(input.asOf, datetime.max.time(), tzinfo=timezone.utc)
-        asof_block_number = BlockNumber.from_timestamp(asof_dt)
-
         debts = self.context.run_model('aave-v2.lending-pool-assets',
                                        input=EmptyInput(),
-                                       return_type=AaveDebtInfos,
-                                       block_number=asof_block_number)
+                                       return_type=AaveDebtInfos)
 
         n_debts = len(debts.aaveDebtInfos)
         positions = []
@@ -62,27 +53,7 @@ class AaveV2GetVAR(Model):
             positions.append(Position(amount=-dbt.totalLiquidity_qty, asset=dbt.token))
         portfolio = Portfolio(positions=positions)
 
-        pls = []
-        pl_assets = set()
-
-        for pos in portfolio:
-            if pos.asset.address not in pl_assets:
-                historical_price_input = HistoricalPriceInput(token=pos.asset,
-                                                              window=input.window,
-                                                              asOf=input.asOf)
-                pl = self.context.run_model(slug='finance.example-historical-price',
-                                            input=json.loads(historical_price_input.json()),
-                                            return_type=PriceList)
-                pls.append(pl)
-                pl_assets.add(pos.asset.address)
-
-        self.logger.info('')
-        var_input = VaRHistoricalInput(
-            portfolio=portfolio,
-            priceLists=pls,
-            interval=input.interval,
-            confidences=input.confidences,
-        )
-        return self.context.run_model(slug='finance.var-engine-historical',
+        var_input = PortfolioVaRInput(portfolio=portfolio, **input)
+        return self.context.run_model(slug='finance.var-portfolio-historical',
                                       input=var_input,
                                       return_type=dict)
